requirements.txt
discord.py==2.3.2
beautifulsoup4==4.12.3
python-dotenv==1.0.0
requests==2.31.0
lxml==5.1.0